pyproject.toml
[project]
name = "z-app-linker"
version = "0.1.0"
description = "links z-app data between them."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "lzstring>=1.0.4",
    "graphviz>=0.20.3",
    "requests>=2.32.5",
]
